+++ python/tests/__snapshots__/test_array_api/test_sklearn_lda[optimized].py
@@ -27,14 +27,13 @@ _NDArray_6 = concat(
 )
 _NDArray_7 = std(_NDArray_6, OptionalIntOrTuple.int(Int(0)))
 _NDArray_7[ndarray_index(std(_NDArray_6, OptionalIntOrTuple.int(Int(0))) == NDArray.scalar(Value.int(Int(0))))] = NDArray.scalar(Value.float(Float(1.0)))
-_TupleNDArray_1 = svd(sqrt(NDArray.scalar(Value.int(NDArray.scalar(Value.float(Float(1.0))).to_int() / Int(147)))) * (_NDArray_6 / _NDArray_7), FALSE)
+_TupleNDArray_1 = svd(sqrt(NDArray.scalar(Value.float(Float(1.0) / Float(147.0)))) * (_NDArray_6 / _NDArray_7), FALSE)
 _Slice_1 = Slice(OptionalInt.none, OptionalInt.some(astype(sum(_TupleNDArray_1[Int(1)] > NDArray.scalar(Value.float(Float(0.0001)))), DType.int32).to_int()))
 _NDArray_8 = (_TupleNDArray_1[Int(2)][IndexKey.multi_axis(MultiAxisIndexKey(MultiAxisIndexKeyItem.slice(_Slice_1)) + _MultiAxisIndexKey_1)] / _NDArray_7).T / _TupleNDArray_1[
     Int(1)
 ][IndexKey.slice(_Slice_1)]
 _TupleNDArray_2 = svd(
-    (sqrt(NDArray.scalar(Value.int((Int(150) * _NDArray_4.to_int()) * (NDArray.scalar(Value.float(Float(1.0))).to_int() / Int(2))))) * (_NDArray_5 - (_NDArray_4 @ _NDArray_5)).T).T
-    @ _NDArray_8,
+    (sqrt((NDArray.scalar(Value.int(Int(150))) * _NDArray_4) * NDArray.scalar(Value.float(Float(1.0) / Float(2.0)))) * (_NDArray_5 - (_NDArray_4 @ _NDArray_5)).T).T @ _NDArray_8,
     FALSE,
 )
 (